// src/api/Customer/endpoints.ts
export enum CustomerApi {
  GetAll = 'Customer/GetAll?Search=',
  GetNames = 'Customer/GetNames',
  GetById = 'Customer/GetById?id=',
  GetWithOrders = 'Customer/GetWithOrders',
  Search = 'Customer/Search',
  SearchResult = 'Customer/SearchResult',
  Add = 'Customer/Add',
  AddAddress = 'Customer/AddAddress',
  Modify = 'Customer/Modify',
  ModifyAddress = 'Customer/ModifyAddress',
  Delete = 'Customer/Delete',
  DeleteAddress = 'Customer/DeleteAddress',
}